Create Capitals!
Checks if your input is in a country in a gigantic CSV file, and if it is it spits out the capital :D 

I will probably make it so I read out the lat and longit as well. It'd be kind of a shame for it to be included in the CSV and not included in the main code.