Now properly finds the largest element in a dictionary.
Now assures that the answer is the lexicographical smallest in the dictionary. Not a fun problem at all. Had to ask a friend on how how to get the lexicographical portion, then they spoiled the whole .find(string, beginning) instead of making a copy of the string and having to worry about replacing it correctly. Would I try it again? 100%.
But its not my favorite problem