Light Sensor! (Really early WIP)
You are reading a digital signal coming from a light sensor connected to a pin on a microcontroller.
You are given an unsorted list of integers from 0-255 representing the light level during a period of time, where each element is the reading for 1 millisecond of time.
Determine the number of milliseconds that a light is being shone.