((Please) (ReMove) (The) (EXTRA))   (better version soon)
A valid parentheses string is either empty (""), "(" + A + ")", or A + B, where A and B are valid parentheses strings, and + represents string concatenation.  For example, "", "()", "(())()", and "(()(()))" are all valid parentheses strings.
Return S after removing the outermost parentheses of every primitive string in the primitive decomposition of S.
Also : I feel there is another slightly better solution. I will most likely do it tomorrow since tomorrow as I am on a roll >:D