Please ReMove The Extra PAREN (Improved!)
Slightly improved solution. Rather than simply through through and appending every item, we simply append a slice of the inner values. Posting the question again for clarity.
A valid parentheses string is either empty (""), "(" + A + ")", or A + B, where A and B are valid parentheses strings, and + represents string concatenation.  For example, "", "()", "(())()", and "(()(()))" are all valid parentheses strings.
Return S after removing the outermost parentheses of every primitive string in the primitive decomposition of S.