Updated documentation!
Everything is same for now, but wanted to update documentation to make the code flow better.

Given an array of integers, find if the array contains any duplicates.

Your function should return true if any value appears at least twice in the array, and it should return false if every element is distinct.