Finally solved! Gonna work on tetris tomorrow. I miss it.
So the problem that was giving me problems (no joke intended) was the following
Text = "a b c d e f g h i j k l m n o p q r s t u v w x y z  "
Due to how  the program was structured we'd get
y z : " " 
even though, that's wonky. So to avoid that, we do a simple check to see if the last character in the text is empty or not. Also realized that "words" is not a great variable name for a variable that is used to get the past 2 indexes. Whoops.