How sentimental are you?
You are given a dictionary of words that indicate what sentiment score they have.
Given a list of strings, output a list of integers that indicates what total sentiment score the strings have.
Not all words are given a sentiment score, so some words or parts of words will not contribute to the sentiment score.
So the example in my code should return

[1.2000000000000002, -1.5, -0.4, 1.3] and it does >:)